UML for MetaXMLParser and MetadataNode.pptx: Created to help focus on the purpose of MetadataNode and MetaXMLParser classes. Created two UML static diagrams for the two classes.
MetadataNode.java: Refactored method names to make more sense.

MetaXMLParser.java: The last four methods in MetaXMLParser are in as placeholders and are not completed. JavaDoc comments have been added to better explain what each method is supposed to do. NEED HELP WITH THESE PLEASE.

TestMetaXMLParser.java: Renamed to TestMetaXMLParser because it is supposed to be a test class. This class should be expanded to also test the methods in the MetadataNode class.