alpha 3.6 - Fixed major issues in MainView and NewSession
Fixed some code that was causing the MainView window to be blank. The code commented out around line 590 will need to be moved to its own method, it should be be in initialize().

NewSession included one deprecated reference to Demo2, fixed and renamed to MainView.
Moved Demo2 to prototypes folder
Created a lib folder for xml template files