Modified grid bag constraints and other GUI polishing
Attempted to modify the grid bag constraints for the JTree in column 1 but it appears like the code will need to be changed in the MyTree class to allow resizing.
Fixed the border padding around the opening message.
Other minor GUI tweaks, like adding a warning message if a session file is invalid (for open menu option).