Metadata Preview Window is fully functional
Debugged the metadata preview JFrame: a new JFrame constructor was being called in the wrong place which set it to a blank frame. Now the frame is set correctly and has a JTextPane inside of a JScrollPane which allows scrolling.
GridBagConstraints were used to make the window resizeable.
Closing the preview window does not close the entire application, which was happening during testing.
WindowListeners were added for the MetadataPreview window for future use.
Within MainView, the preview window call was set inside a new EventQueue runnable.
The textbox previously named openingScreen was renamed to openingMsg.