Changed open functionality for XMLs in MainView
The open file menu will now create a string if the file was an XML file and then it will pass that string to the MetadataPreview class to display a text preview of the XML file. The previous functionality was to open the XML file in the system's default editor.
MetadataPreview now scrolls to the top of the text output before being displayed to the user and there exists a way to set a custom title bar message.
FileOps1 is in the process of being overhauled to become more useful.