New contours template, minor GUI tweaks
Attempted to get the USGSFiles dialog to be centered on the screen by setting the location relative to its parent, which will always be the NewSession frame - but it has no effect.
Tried to start implementing action on the checkboxes for USGS files but had issues with paths to the xml resources.