Fixed major syntax errors in the Contours XML template
There were many instances of double dashes "--" appearing in the comments, which is not a good practice and was causing the parser to throw an error.

Added clear user facing error handling should an error occur during the DOM parsing or the setup of the metadata node tree. An error dialog will popup with similar messages to what the terminal would display.

Removed obsolete and deprecated class FileTypeFilter