Version alpha 5.2 - JTree size improved and USGSFiles checkboxes functional
Changed the size of the JTree for the main tree navigation, before it was too narrrow.
Greatly improved the SharedData class, it now allows multiple template files by using an ArrayList of Strings for the absolute file paths.
In the test environment, got the checkboxes to get the template XML files from the classpath resources, but it is not working in the JAR file.
Disabled checkboxes which are not yet implemented.