alpha 5.3 - USGS templates work in JAR
The classified point cloud checkbox incorporates the correct and neccessary code to use the included USGS xml templates files within the jar.
The main process is that the file within the jar must be read by an input stream and then sent to a new file via a file output stream.
The code currently places the xml file in a directory named 'lib' that is in the same directory the jar file is located in.

Currently selecting multiple checkboxes does not work and not all checkboxes include the needed actions.
Overall, this is a major improvement compared to version alpha 5.3