alpha 6.0 - USGS Files fully working with radio buttons, icon images throughout
Completely redesigned the USGS Files class to use radio buttons instead of checkboxes, which forces the user to only pick one of the templates.
Used a helper method to get files from the classpath and set the template files. This allows the program to work correctly when packaged into a JAR file.

Redesigned many classes to include an image icon, since I was able to read files from the classpath. This is implemented in every JFrame and JDialog called from the main GUI. The use of parent components makes this easier.
The welcome and about dialogs now have 'branding' (icons).

Only the save and export functions of the current program are not fully functional