Save feature finished, reliability/debugging improvements
Finished the save feature, with added code in FileOps1 to add a .xsm extension if the user specified file does not include that already.

Made debugging console output messages more clear in MainView and NewSession.

Handled a potential NullPointerException in USGSFiles which would occur when a user clicks the OK button without selecting anything. Now the action listener checks for that and displays a dialog to the user if they did not select any radio buttons (templates).