Some progress made on export
Changed the code to check for null pointers when doing an export, before it only checked for the end of the array - even though the array might not have been full.
Prompted the user to enter a project number, which is saved as a string.

Still need to work on getting the program to save the exported files in the right location.